add authorization

+++ routes/campgrounds.js
@@ -6,7 +6,8 @@ const Campground = require('../models/campground');
 const Review = require('../models/review');
 const { campgroundSchema, reviewSchema } = require('../schemas.js');
 const Joi = require('joi');
-const isLoggedIn = require('../middleware');
+const isLoggedIn = require('../middlewares/authentication');
+const {isAuthor } = require('../middlewares/authorazition');
 
 const validateCampground = (req, res, next) => {
     const { error } = campgroundSchema.validate(req.body);
@@ -52,6 +53,7 @@ router.get('/new', isLoggedIn, (req, res) => {
 router.post('/', isLoggedIn, validateCampground, catchAsync(async (req, res, next) => {
     // if(!req.body.campground) throw new ExpressError('Invalid Campground Data', 400);
     const campground = new Campground(req.body.campground);
+    campground.author = req.user._id;
     await campground.save();
     req.flash('success', 'successfully made a new campground');
     res.redirect(`/campgrounds/${campground._id}`);
@@ -59,7 +61,14 @@ router.post('/', isLoggedIn, validateCampground, catchAsync(async (req, res, nex
 }))
 
 router.get('/:id', isLoggedIn, async (req, res) => {
-    const campground = await Campground.findById(req.params.id).populate('reviews');
+    const campground = await Campground.findById(req.params.id).populate({
+        path: 'reviews',
+        populate: {
+            path: 'author',
+        },
+        model: 'Review'
+    }).populate('author');
+    console.log(campground);
     if (!campground) {
         req.flash('error', 'cannot find that campground');
         return res.redirect('/campgrounds');
@@ -67,22 +76,32 @@ router.get('/:id', isLoggedIn, async (req, res) => {
     res.render('campgrounds/show', { campground });
 })
 
-router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
     const campground = await Campground.findById(req.params.id);
+    if (!campground) {
+        req.flash('error', 'cannot find that campground');
+        return res.redirect('/campgrounds');
+    }
     res.render('campgrounds/edit', { campground });
 }))
 
 
 
-router.put('/:id', isLoggedIn, catchAsync(async (req, res) => {
+router.put('/:id', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
     const { id } = req.params;
     const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
     req.flash('success', 'successfully updated a campground');
     res.redirect(`/campgrounds/${campground._id}`);
 }))
 
-router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
+router.delete('/:id', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    const currentUserId = JSON.stringify(req.user._id);
+    if (campIdStr !== currentUserId) {
+        req.flash('error', 'you do not have permission to do that');
+        return res.redirect(`/campgrounds/${campground._id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'successfully deleted a campground');
     res.redirect('/campgrounds');
